AVA/requirements.txt
python-dotenv
PyQt5
pyttsx4
pygame
requests
simpleaudio
SpeechRecognition
playsound
python-dateutil
bcrypt
opencv-python
pillow
openai
google-genai
pycaw
comtypes
tiktoken
SkillsManager
SynLrn
SynMem
BitSig
MediaCapture
